polish useDocs api surface
